change type

+++ app/src/main/java/com/wisemotions/pocfirestorekotlin/MainActivity.kt
@@ -14,6 +14,6 @@ class MainActivity : AppCompatActivity() {
 
         val db = FirebaseFirestore.getInstance()
 
-        val pippo:Int
+        val pippo:Float
     }
 }
